move writeMetrics to own module
